Speed up epsilon neighborhood computation in DBSCAN
DBSCAN needs the compute epsilon neighboroods (fixed-distance nearest neighbors);
this is the bottleneck of the algorithm. It can be sped up by pre-computing the
epsilon neighborhoods and storing the information. This has the upside of making
the epsilon queries constant time for any epsilon, but has the downside of being
memory-intensive; a future alternative may be to approximate using an integer
lattice first to save memory, at the cost of making the epsilon queries slower
than constant time and proportional to the size of epsilon.